requirements.txt
python-dotenv
openai
gspread
google-auth
flask
gunicorn
pandas
openpyxl
